Analysis almost done, codes as of Oct-11
Finished the work on the logn-logs and sensitivity. This version is after the new run of wavdetect with no background map but with exposure and psfmaps has been done. There are hints of an increased spurious fraction in the hard band (logn-logs requires a cut in probability from 10-4 to 5e-5 to get reasonable results).